requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.9
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
